Added pricing table and testimonial snippet
Colors still need to be adjusted, testimonials still need to be filled out, and there is a big white stripe at the top of each section. I'm not sure what that is or how to get rid of it.
I also messed up the header and I didn't even touch any code near the header...